optimisation for getTexture() texture uses RGBA pixels, not RGB. in which case, RGB pixels are not calculated unless needed. this increases the performance by 2-5 fps.
also added a method for enabling texture cache - setTextureCache()
i still haven't been able to weak link the CoreVideo framework, so it still crashes when this in enabled on any devices not running iOS5.
it is up to the user to make sure they are only deploying for iOS5.
by default it is set to false.